% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maxquant.R
\name{add_label_column}
\alias{add_label_column}
\title{Add a label column to the data frame subset for the volcano plot}
\usage{
add_label_column(df_subset, label_points)
}
\arguments{
\item{df_subset}{the data frame to be used for the volcano plot, which should
contain the columns "Gene.names", "Protein.names", and "Protein.IDs" and a
column "vp_colorcode" which contains the colour code for each point in the
volcano plot (0, 1, 2, 3, 4, or 5)}

\item{label_points}{character string specifying which points to label in the
volcano plot. This can be one of the following: "none", "all", a character
vector of selected proteins to label, a code for auto-labelling}
}
\value{
the input data frame with an additional column "point_labels" containing the
labels for the points to be labelled in the volcano plot. The "point_labels"
}
\description{
Add a label column to the data frame subset for the volcano plot
}
\keyword{internal}
