% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_volcano.R
\name{create_volcano}
\alias{create_volcano}
\title{Build a volcano plot with create_volcano()}
\usage{
create_volcano(dataPath, bioSig, statSig)
}
\arguments{
\item{dataPath}{A file path to locate your data}

\item{bioSig}{A numeric value representing biological significant}

\item{statSig}{A numeric value representing statistical significant}
}
\value{
This function returns a volcano plot
}
\description{
This function can be used to build a volcano plot from Kallisto outputs to visualize your genomic or transcriptomic data.
}
