% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maxquant.R
\name{determine_groups_from_df_subset}
\alias{determine_groups_from_df_subset}
\title{Determine the experimental groups in a data frame subset}
\usage{
determine_groups_from_df_subset(df_subset, meas)
}
\arguments{
\item{df_subset}{a data frame containing the subset of the MaxQuant data that will be used to
generate the volcano plot.}

\item{meas}{string indicating the measurement type}
}
\value{
vector of two strings containing the names of the experimental groups
}
\description{
This function takes a data frame containing a subset of the MaxQuant data and
determines the experimental groups that are present in the data frame by looking
at the column names.
}
\keyword{internal}
