% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_maxquant}
\alias{load_maxquant}
\title{Load MaxQuant output file}
\usage{
load_maxquant(file = "proteinGroups.txt", datadir = "Data", clean = TRUE)
}
\arguments{
\item{file}{character string specifying the name of the MaxQuant output file
to load (default is "proteinGroups.txt").}

\item{datadir}{character string specifying the directory where the MaxQuant
output file is located (default is "Data").}

\item{clean}{boolean indicating whether to perform cleaning steps on the data
(default is TRUE). If TRUE, rows with a "+" in the
"Only.identified.by.site", "Reverse", or "Potential.contaminant" columns
will be removed, and missing gene and protein names will be filled in based
on the "Fasta.headers" column.}
}
\value{
A data frame containing the MaxQuant data.
}
\description{
Load a \code{proteinGroups.txt} file from MaxQuant and perform some cleaning steps
to prepare the data for analysis. Proteins that are only identified by site,
reverse hits, and potential contaminants are removed, and missing gene and
protein names are filled in based on the "Fasta.headers" column. Note: in the
case of multiple identified proteins, the first one in the list is kept.
}
