% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_multiple_maxquant.R
\name{load_multiple_maxquant}
\alias{load_multiple_maxquant}
\title{Load multiple MaxQuant output files}
\usage{
load_multiple_maxquant(
  file = "proteinGroups.txt",
  datadir = "Data",
  clean = TRUE
)
}
\arguments{
\item{file}{character string specifying the name of the MaxQuant output file
to load (default is "proteinGroups.txt").}

\item{datadir}{character string specifying the directory where the MaxQuant
output files are located (default is "Data"). The function performs a
recursive search in this directory for all files matching the specified
\code{file} name.}

\item{clean}{boolean indicating whether to perform cleaning steps on the data
(default is TRUE). If TRUE, rows with a "+" in the
"Only.identified.by.site", "Reverse", or "Potential.contaminant" columns
will be removed, and missing gene and protein names will be filled in based
on the "Fasta.headers" column.}
}
\value{
A list of data frames containing the MaxQuant data.
}
\description{
Load several \code{proteinGroups.txt} files from MaxQuant and perform some
cleaning steps to prepare the data for combined analysis. If only one file is
to be analyzed, use \code{load_maxquant()} instead. For automatic loading, place
\code{proteinGroups.txt} files in subfolders within \code{datadir}. Proteins that are
only identified by site, reverse hits, and potential contaminants are
removed, and missing gene and protein names are filled in based on the
"Fasta.headers" column. Note: in the case of multiple identified proteins,
the first one in the list is kept.
}
\examples{
\dontrun{
load_multiple_maxquant()
}
}
