% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maxquant.R
\name{validate_df_subset}
\alias{validate_df_subset}
\title{Validate a data frame to generate a volcano plot}
\usage{
validate_df_subset(df_subset)
}
\arguments{
\item{df_subset}{data frame containing the subset of the MaxQuant data that
will be used to generate the volcano plot. This data frame should contain
the columns "Protein.IDs", "Gene.names", "Protein.names", "meas.ratio",
"p.value", and "neg.log10.p.value". The "Gene.names" and "Protein.names"
columns are optional, but if they are not present, a warning will be
printed.}
}
\value{
A boolean indicating whether the data frame contains the required
columns to generate a volcano plot, and a warning if the optional columns
are not present.
}
\description{
Checks if we have all the necessary components to generate a volcano plot.
}
\keyword{internal}
