% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot_maxquant.R
\name{volcano_plot_maxquant}
\alias{volcano_plot_maxquant}
\title{Volcano Plot of Maxquant Data}
\usage{
volcano_plot_maxquant(
  df_subset = NULL,
  meas = "LFQ.intensity",
  threshold_p = 0.05,
  threshold_fc = 1,
  vp_colours = NULL,
  fsize = 8
)
}
\arguments{
\item{df_subset}{a data frame containing the subset of the MaxQuant data that
will be used to generate the plot}

\item{meas}{string indicating the measurement type (e.g. "LFQ.intensity",
"Intensity", etc.)}

\item{threshold_p}{numeric indicating the p-value threshold}

\item{threshold_fc}{numeric indicating the fold change threshold (in log2
space, i.e. 1 is a 2-fold change, 2 is a 4-fold change, etc.)}

\item{vp_colours}{a named vector of colours for the volcano plot. The names
of the vector should be the integers 0 through 5, which correspond to the
different combinations of p-value and fold change thresholds. If NULL,
default colours will be used.}

\item{fsize}{numeric indicating the font size}
}
\value{
ggplot object containing the volcano plot
}
\description{
This function generates a volcano plot from a subset of MaxQuant data. The
data frame should contain the following columns: Protein.IDs, Gene.names,
Protein.names, meas.ratio, p.value, and neg.log10.p.value. The function will
add a colour code column to the data frame based on the p-value and fold
change thresholds, and then use ggplot2 to create the volcano plot.
}
