% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_maxquant.R
\name{workflow_maxquant}
\alias{workflow_maxquant}
\title{Workflow for MaxQuant data analysis}
\usage{
workflow_maxquant()
}
\value{
A ggplot object containing the volcano plot generated from the MaxQuant data.
}
\description{
The purpose of this function is to provide a single function that will take a
\code{proteinGroups.txt} produced by MaxQuant, perform the necessary data
processing, and generate a volcano plot. This function will call the
\code{load_maxquant()}, \code{process_maxquant()}, and \code{volcano_plot_maxquant()}
functions in turn. The user will be asked to specify which groups are group 1 or group 2.
}
\examples{
\dontrun{
workflow_maxquant()
}
}
